+++ circus-train-avro/src/main/java/com/hotels/bdp/circustrain/avro/transformation/AbstractAvroSerDeTransformation.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,18 @@
  */
 package com.hotels.bdp.circustrain.avro.transformation;
 
+import static org.apache.commons.lang3.StringUtils.isNotEmpty;
+
 import static com.hotels.bdp.circustrain.avro.util.AvroStringUtils.argsPresent;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.Collections;
 import java.util.Map;
 
 import org.apache.commons.lang3.StringUtils;
 
+import com.hotels.bdp.circustrain.api.CircusTrainException;
 import com.hotels.bdp.circustrain.api.event.EventTableReplication;
 import com.hotels.bdp.circustrain.api.event.TableReplicationListener;
 import com.hotels.bdp.circustrain.avro.conf.AvroSerDeConfig;
@@ -48,6 +53,28 @@ public abstract class AbstractAvroSerDeTransformation implements TableReplicatio
 
   protected boolean avroTransformationSpecified() {
     return argsPresent(getAvroSchemaDestinationFolder(), getEventId());
+  }
+
+  protected String constructSource(String source) {
+    String scheme = null;
+    Object overrideScheme = avroSerdeConfigOverride
+        .get(AvroSerDeConfig.TABLE_REPLICATION_OVERRIDE_SOURCE_AVRO_URL_SCHEME);
+    if (overrideScheme != null && StringUtils.isNotBlank(overrideScheme.toString())) {
+      scheme = overrideScheme.toString();
+    } else {
+      scheme = avroSerDeConfig.getSourceAvroUrlScheme();
+    }
+
+    if (isNotEmpty(scheme)) {
+      try {
+        URI uri = new URI(source);
+        source = new URI(scheme, uri.getUserInfo(), uri.getHost(), uri.getPort(), uri.getPath(), uri.getQuery(),
+            uri.getFragment()).toString();
+      } catch (URISyntaxException e) {
+        throw new CircusTrainException(e);
+      }
+    }
+    return source;
   }
 
   protected String getAvroSchemaDestinationFolder() {