+++ circus-train-avro/src/main/java/com/hotels/bdp/circustrain/avro/transformation/AvroSerDeTableTransformation.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -50,7 +50,7 @@ public final class AvroSerDeTableTransformation extends AbstractAvroSerDeTransfo
   }
 
   private Table apply(Table table, String avroSchemaDestination) {
-    String source = HiveObjectUtils.getParameter(table, AVRO_SCHEMA_URL_PARAMETER);
+    String source = constructSource(HiveObjectUtils.getParameter(table, AVRO_SCHEMA_URL_PARAMETER));
     if (argsPresent(source, avroSchemaDestination)) {
       String destinationPath = copier.copy(source, avroSchemaDestination).toString();
       HiveObjectUtils.updateSerDeUrl(table, AVRO_SCHEMA_URL_PARAMETER, destinationPath);